feat: add feature manager generator
