--- analysis_options.yaml
@@ -1,206 +1,0 @@
-analyzer:
-  language:
-    strict-casts: true
-    strict-inference: true
-    strict-raw-types: true
-  errors:
-    require_trailing_commas: info
-    missing_required_param: error
-    missing_return: error
-  exclude:
-    - 'bin/cache/**'
-    - 'lib/generated/**'
-    - 'lib/**/*.g.dart'
-    - 'test/**/*.mocks.dart'
-
-# Definitions: https://dart-lang.github.io/linter/lints/
-linter:
-  rules:
-    # Error Rules
-    - always_use_package_imports
-    - avoid_empty_else
-    - avoid_print
-    - avoid_relative_lib_imports
-    - avoid_returning_null_for_future
-    - avoid_slow_async_io
-    - avoid_types_as_parameter_names
-    - avoid_web_libraries_in_flutter
-    - cancel_subscriptions
-    - close_sinks
-    - comment_references
-    - control_flow_in_finally
-    - empty_statements
-    - hash_and_equals
-    - literal_only_boolean_expressions
-    - no_adjacent_strings_in_list
-    - no_duplicate_case_values
-    - no_logic_in_create_state
-    - prefer_void_to_null
-    - test_types_in_equals
-    - throw_in_finally
-    - unnecessary_statements
-    - unrelated_type_equality_checks
-    - use_key_in_widget_constructors
-    - valid_regexps
-    - null_check_on_nullable_type_parameter
-    - collection_methods_unrelated_type
-
-    # Style Rules
-    - always_declare_return_types
-    - always_put_control_body_on_new_line
-    - always_put_required_named_parameters_first
-    - always_require_non_null_named_parameters
-    - annotate_overrides
-    - avoid_bool_literals_in_conditional_expressions
-    - avoid_catching_errors
-    - avoid_double_and_int_checks
-    - avoid_equals_and_hash_code_on_mutable_classes
-    - avoid_escaping_inner_quotes
-    - avoid_field_initializers_in_const_classes
-    - avoid_function_literals_in_foreach_calls
-    - avoid_implementing_value_types
-    - avoid_init_to_null
-    - avoid_js_rounded_ints
-    - avoid_null_checks_in_equality_operators
-    - avoid_private_typedef_functions
-    - avoid_renaming_method_parameters
-    - avoid_return_types_on_setters
-    - avoid_returning_null
-    - avoid_returning_null_for_void
-    - avoid_returning_this
-    - avoid_setters_without_getters
-    - avoid_shadowing_type_parameters
-    - avoid_single_cascade_in_expression_statements
-    - avoid_types_on_closure_parameters
-    - avoid_unnecessary_containers
-    - avoid_unused_constructor_parameters
-    - avoid_void_async
-    - await_only_futures
-    - camel_case_extensions
-    - camel_case_types
-    - constant_identifier_names
-    - curly_braces_in_flow_control_structures
-    - directives_ordering
-    - empty_catches
-    - empty_constructor_bodies
-    - exhaustive_cases
-    - file_names
-    - flutter_style_todos
-    - implementation_imports
-    - join_return_with_assignment
-    - leading_newlines_in_multiline_strings
-    - library_names
-    - library_prefixes
-    - missing_whitespace_between_adjacent_strings
-    - no_default_cases
-    - no_runtimeType_toString
-    - non_constant_identifier_names
-    - no_leading_underscores_for_library_prefixes
-    - no_leading_underscores_for_local_identifiers
-    - null_closures
-    - omit_local_variable_types
-    - only_throw_errors
-    - overridden_fields
-    - package_api_docs
-    - package_prefixed_library_names
-    - parameter_assignments
-    - prefer_adjacent_string_concatenation
-    - prefer_asserts_in_initializer_lists
-    - prefer_collection_literals
-    - prefer_conditional_assignment
-    - prefer_const_constructors
-    - prefer_const_constructors_in_immutables
-    - prefer_const_declarations
-    - prefer_const_literals_to_create_immutables
-    - prefer_contains
-    - prefer_final_fields
-    - prefer_final_in_for_each
-    - prefer_final_locals
-    - prefer_for_elements_to_map_fromIterable
-    - prefer_foreach
-    - prefer_function_declarations_over_variables
-    - prefer_generic_function_type_aliases
-    - prefer_if_elements_to_conditional_expressions
-    - prefer_if_null_operators
-    - prefer_initializing_formals
-    - prefer_inlined_adds
-    - prefer_interpolation_to_compose_strings
-    - prefer_int_literals
-    - prefer_is_empty
-    - prefer_is_not_empty
-    - prefer_is_not_operator
-    - prefer_iterable_whereType
-    - prefer_mixin
-    - prefer_null_aware_operators
-    - prefer_single_quotes
-    - prefer_spread_collections
-    - prefer_typing_uninitialized_variables
-    - provide_deprecation_message
-    - recursive_getters
-    - require_trailing_commas
-    - sized_box_for_whitespace
-    - sized_box_shrink_expand
-    - slash_for_doc_comments
-    - sort_child_properties_last
-    - sort_constructors_first
-    - sort_unnamed_constructors_first
-    - type_init_formals
-    - tighten_type_of_initializing_formals
-    - unawaited_futures
-    - unnecessary_await_in_return
-    - unnecessary_brace_in_string_interps
-    - unnecessary_const
-    - unnecessary_constructor_name
-    - unnecessary_getters_setters
-    - unnecessary_late
-    - unnecessary_lambdas
-    - unnecessary_new
-    - unnecessary_null_aware_assignments
-    - unnecessary_null_in_if_null_operators
-    - unnecessary_nullable_for_final_variable_declarations
-    - unnecessary_overrides
-    - unnecessary_parenthesis
-    - unnecessary_raw_strings
-    - unnecessary_string_escapes
-    - unnecessary_string_interpolations
-    - unnecessary_this
-    - unnecessary_breaks
-    - unnecessary_to_list_in_spreads
-    - invalid_case_patterns
-    - type_literal_in_constant_pattern
-    - use_full_hex_values_for_flutter_colors
-    - use_function_type_syntax_for_parameters
-    - use_is_even_rather_than_modulo
-    - use_late_for_private_fields_and_variables
-    - use_raw_strings
-    - use_rethrow_when_possible
-    - use_if_null_to_convert_nulls_to_bools
-    - use_string_buffers
-    - use_super_parameters
-    - use_to_and_as_if_applicable
-    - use_enums
-    - use_decorated_box
-    - use_colored_box
-    - void_checks
-    - combinators_ordering
-    - unreachable_from_main
-    - implicit_call_tearoffs
-    - enable_null_safety
-
-    # Pub Rules
-    - package_names
-    - depend_on_referenced_packages
-
-    # Ignored Rules (intentionally)
-    #- discarded_futures
-    #- dangling_library_doc_comments
-    #- unnecessary_library_directive
-    #- library_annotations
-    #- use_string_in_part_of_directives
-    #- one_member_abstracts
-    #- prefer_constructors_over_static_methods
-    #- use_setters_to_change_properties
-    #- library_private_types_in_public_api
-    #- avoid_positional_boolean_parameters
-    #- avoid_classes_with_only_static_members
-    #- avoid_catches_without_on_clauses
