+++ example/lib/features.dart
@@ -1,51 +1,63 @@
+import 'package:feature_manager/annotations.dart';
+import 'package:feature_manager/feature.dart';
 import 'package:feature_manager/feature_manager.dart';
 
-class Features {
-  static const Feature textFeature = Feature(
+part 'features.fm.dart';
+
+@FeatureManagerInit()
+class AppFeatures {
+  AppFeatures({
+    required this.textFeature,
+    required this.booleanFeature,
+    required this.doubleFeature,
+    required this.integerFeature,
+    required this.jsonFeature,
+  });
+
+  factory AppFeatures.instance() => _$AppFeatures();
+
+  @FeatureOptions(
     key: 'dev-prefs-text-pref',
     title: 'Text pref',
     description: 'This is text preference',
     defaultValue: '',
     valueType: FeatureValueType.text,
-  );
+  )
+  final TextFeature textFeature;
 
-  static const Feature booleanFeature = Feature(
+  @FeatureOptions(
     key: 'dev-prefs-bool-pref',
     title: 'Toggle pref',
     description: 'This is toggle preference',
     defaultValue: false,
     valueType: FeatureValueType.toggle,
-  );
+  )
+  final BooleanFeature booleanFeature;
 
-  static const Feature doubleFeature = Feature(
+  @FeatureOptions(
     key: 'dev-prefs-double-pref',
     title: 'Number double pref',
     description: 'This is number double preference',
-    defaultValue: 0.0,
+    defaultValue: null,
     valueType: FeatureValueType.doubleNumber,
-  );
+  )
+  final DoubleFeature doubleFeature;
 
-  static const Feature integerFeature = Feature(
+  @FeatureOptions(
     key: 'dev-prefs-integer-pref',
     title: 'Number integer pref',
     description: 'This is number integer preference',
-    defaultValue: 0,
+    defaultValue: null,
     valueType: FeatureValueType.integerNumber,
-  );
+  )
+  final IntegerFeature integerFeature;
 
-  static const Feature jsonFeature = Feature(
+  @FeatureOptions(
     key: 'dev-prefs-json-pref',
     title: 'Json pref',
     description: 'This is json preference',
     defaultValue: "{value: 'Json default value'}",
     valueType: FeatureValueType.json,
-  );
-
-  static const List<Feature> values = <Feature>[
-    Features.textFeature,
-    Features.booleanFeature,
-    Features.doubleFeature,
-    Features.integerFeature,
-    Features.jsonFeature,
-  ];
+  )
+  final Feature jsonFeature;
 }
