+++ example/pubspec.yaml
@@ -3,26 +3,36 @@ description: Example project for feature manager
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.2.4 <4.0.0'
-  flutter: '>=3.16.0'
+  sdk: ^3.5.3
 
 dependencies:
   flutter:
     sdk: flutter
 
   feature_manager:
-    path: ../
+    # git:
+    #   url: git@github.com:hrubalski/feature-manager.git
+    #   ref: feat-feature-manager-generator
+    #   path: feature_manager
+    path: ../feature_manager
 
   provider: ^6.1.2
-  shared_preferences: ^2.2.2
+  shared_preferences: ^2.3.2
 
 dev_dependencies:
+  build_runner:
   flutter_test:
     sdk: flutter
+  feature_manager_generator:
+    # git:
+    #   url: git@github.com:hrubalski/feature-manager.git
+    #   ref: feat-feature-manager-generator
+    #   path: generator
+    path: ../generator/
 
 # The following section is specific to Flutter.
 flutter:
