+++ feature_manager/lib/annotations.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/annotations/annotations.dart';
