+++ feature_manager/lib/feature.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/models/feature.dart';
