+++ feature_manager/lib/feature_manager.dart
@@ -1,0 +1,4 @@
+library;
+
+export 'src/feature_manager.dart';
+export 'src/presentation/features_screen.dart';
