+++ feature_manager/lib/src/cubit/cubit.dart
