+++ feature_manager/pubspec.yaml
@@ -1,0 +1,20 @@
+# Analyze before publish: flutter pub publish --dry-run
+# Publish: flutter pub publish
+name: feature_manager
+description: Feature manager for developer preferences and experiments
+version: 2.0.0
+repository: https://github.com/theRealGetman/feature-manager/tree/master/feature_manager
+issue_tracker: https://github.com/theRealGetman/feature-manager/issues
+
+environment:
+  sdk: ">=3.2.4 <4.0.0"
+
+dependencies:
+  shared_preferences: ">=2.1.2 <3.0.0"
+  flutter:
+    sdk: flutter
+
+dev_dependencies:
+  lints: ^5.0.0
+  flutter_test:
+    sdk: flutter
