+++ generator/lib/feature_manager_generator.dart
@@ -1,0 +1,6 @@
+library;
+
+import 'package:build/build.dart';
+import 'package:feature_manager_generator/src/factory.dart';
+
+Builder featureManagerGenerator(BuilderOptions options) => generatorFactoryBuilder(options);
