+++ generator/pubspec.yaml
@@ -1,0 +1,20 @@
+name: feature_manager_generator
+description: Feature manager generator for developer preferences and experiments
+repository: https://github.com/theRealGetman/feature-manager/tree/master/generator
+issue_tracker: https://github.com/theRealGetman/feature-manager/issues
+
+version: 2.0.0
+
+environment:
+  sdk: ">=3.2.4 <4.0.0"
+
+# Add regular dependencies here.
+dependencies:
+  analyzer: ^6.7.0
+  build: ^2.4.1
+  source_gen: ^1.5.0
+  feature_manager: ^2.0.0
+
+dev_dependencies:
+  lints: ^5.0.0
+  test: ^1.25.0
