+++ generator/test/feature_manager_generator_test.dart
@@ -1,0 +1,1 @@
+void main() {}
