--- lib/feature_manager.dart
@@ -1,5 +1,0 @@
-library feature_manager;
-
-export 'src/domain/models/feature.dart';
-export 'src/presentation/features_screen.dart';
-export 'src/utils/feature_manager.dart';
