--- pubspec.yaml
@@ -1,26 +1,0 @@
-# Analyze before publish: flutter pub publish --dry-run
-# Publish: flutter pub publish
-name: feature_manager
-description: Feature manager for developer preferences and experiments
-version: 2.0.0
-repository: https://github.com/theRealGetman/feature-manager
-issue_tracker: https://github.com/theRealGetman/feature-manager/issues
-
-environment:
-  sdk: '>=3.2.4 <4.0.0'
-  flutter: '>=3.16.0'
-
-dependencies:
-  flutter:
-    sdk: flutter
-
-  shared_preferences: '>=2.1.2 <3.0.0'
-
-dev_dependencies:
-  flutter_test:
-    sdk: flutter
-
-  build_runner: ^2.4.8
-
-# The following section is specific to Flutter.
-flutter:
